JMH benchmarks moved to platform
